+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bubblescript",
-  "version": "0.0.12",
+  "version": "0.0.13",
   "scripts": {
     "test": "mocha ./test/**/*.js"
   },
